assignments/assignment 4/requirements.txt
facenet_pytorch==2.6.0
matplotlib==3.8.4
pandas==2.2.2
Pillow==10.3.0
torch==2.3.0
